Move RequestHandle test cases to separate class
